Eliminar Pass mejorado. Login con contraseña obligatoria en php. Estilos corregidos

+++ bbdd/passwords.php
@@ -64,10 +64,10 @@ class Password extends Conexion {
             $passid = $row["password_id"];
 
             echo '<div id="savedpass" class="passwords__passdiv">
-                    <h4 class="passdiv__nombre" id="' . $passid . '">' . $nombrePass . '</h4>
-                    <a class="passdiv__password" id="' . $password . '">' . $password . '</a>
+                    <h4 class="passdiv__nombre">' . $nombrePass . '</h4>
+                    <a class="passdiv__password">' . $password . '</a>
                     <nav class="passdiv__nav">
-                    <a class="nav__eliminar" id="deleteClick" onclick="eliminarPass(\'' . $password . '\', \'' . $passid . '\')">Eliminar</a>
+                    <a class="nav__eliminar" id="deleteClick" onclick="eliminarPass(\'' . $passid . '\')">Eliminar</a>
                     <a class="nav__copiar" onclick="copiar(\'' . $password . '\')">Copiar</a>
                     </nav>
                 </div>';
@@ -90,9 +90,10 @@ class Password extends Conexion {
         }
     }
 
-    public function deletePass($id){
-        $stmt = $this->getPdo()->prepare("DELETE FROM saved_passwords WHERE password_id = :id");
+    public function deletePass($id, $idUsuario){
+        $stmt = $this->getPdo()->prepare("DELETE FROM saved_passwords WHERE password_id = :id AND user_id = :idUsuario");
         $stmt->bindParam(':id', $id);
+        $stmt->bindParam(':idUsuario', $idUsuario);
         $stmt->execute();
     }
 }
